Added flag --keepfiles which prevents deleting of the test directories in case of failure

+++ tests/regression/regression.py
@@ -17,6 +17,8 @@ _THIS_DIR = os.path.abspath(os.path.dirname(__file__))
 _TEST_REGEXP = "^test_.*$"
 _PYTHON = sys.executable
 
+_KEEP_FAILED_OUTPUTS = False
+_A_TEST_FAILED       = False
 
 class TestCase(namedtuple(
         "TestCase",
@@ -45,6 +47,7 @@ class TestCase(namedtuple(
 def launch_test_set(test_cases, repo_path, tag_regexp=_TEST_REGEXP):
     """
     """
+    _A_TEST_FAILED = False
     _print_sep("Test session starts")
     test_tag = find_tag(repo_path, tag_regexp)
     print("Testing against tag \"{tag}\":\n"
@@ -55,6 +58,8 @@ def launch_test_set(test_cases, repo_path, tag_regexp=_TEST_REGEXP):
         print("Cloning repository into {}".format(new_repo_path))
         clone_revision(repo_path, test_tag.commit.hexsha, new_repo_path)
         result = find_regressions(test_cases, new_repo_path, repo_path)
+        if not result:
+            _A_TEST_FAILED = True
     _print_sep("Regression finished in {:.2f}s".format(regression_timer.get_duration()))
     if not result:
         raise RegressionTestFailed()
@@ -109,6 +114,10 @@ def run_test_case(test_case, valid_path, test_path):
     test_script = os.path.join(test_path, test_case.script)
     valid_outpath = os.path.join(valid_path, test_case.output)
     test_outpath = os.path.join(test_path, test_case.output)
+    
+    anEx = None
+    fail = None
+    
     try:
         with Timer() as test_timer:
             if test_case.pre_hook:
@@ -118,9 +127,30 @@ def run_test_case(test_case, valid_path, test_path):
             test_result = _launch_command(test_path, test_script, test_case.arguments)
         result = TestResult(test_case, valid_result, test_result,
                             valid_outpath, test_outpath, test_timer.get_duration())
+        
+        fail = not result.is_success
+        
+    except Exception as inEx:
+        anEx = inEx
+        fail = True
     finally:
+        print ("  Done",end="")
+    
+        
+    if _KEEP_FAILED_OUTPUTS and fail :
+        print ("")
+        print ("Test failed, keeping output files: ")
+        print ("  	valid_outpath = < %s > " % valid_outpath)
+        print ("  	test_outpath  = < %s > " % test_outpath)
+    
+    else:
         _remove_if_exists(valid_outpath)
         _remove_if_exists(test_outpath)
+    
+    if anEx:
+        raise anEx
+      
+
     return result
 
 
@@ -318,7 +348,8 @@ def _temporary_dir():
         dir_path = tempfile.mkdtemp()
         yield dir_path
     finally:
-        shutil.rmtree(dir_path)
+        if not (_A_TEST_FAILED and _KEEP_FAILED_OUTPUTS): 
+            shutil.rmtree(dir_path)
 
 
 class Timer(object):
